Dispatch canvas related helpers to helpers.canvas
