+++ samples/charts/bar/horizontal.html
@@ -25,7 +25,7 @@
 	<button id="removeData">Remove Data</button>
 	<script>
 		var MONTHS = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'];
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		var horizontalBarChartData = {
 			labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
 			datasets: [{
