+++ samples/charts/bubble.html
@@ -27,7 +27,7 @@
 		var DEFAULT_DATASET_SIZE = 7;
 
 		var addedCount = 0;
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		var bubbleChartData = {
 			animation: {
 				duration: 10000
