+++ samples/charts/polar-area.html
@@ -27,7 +27,7 @@
 		};
 
 		var chartColors = window.chartColors;
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		var config = {
 			data: {
 				datasets: [{
