+++ samples/charts/radar-skip-points.html
@@ -24,7 +24,7 @@
 			return Math.round(Math.random() * 100);
 		};
 
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		var config = {
 			type: 'radar',
 			data: {
