+++ samples/charts/scatter/basic.html
@@ -20,7 +20,7 @@
 	</div>
 	<button id="randomizeData">Randomize Data</button>
 	<script>
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		function generateData() {
 			var data = [];
 			for (var i = 0; i < 7; i++) {
