+++ samples/charts/scatter/multi-axis.html
@@ -20,7 +20,7 @@
 	</div>
 	<button id="randomizeData">Randomize Data</button>
 	<script>
-	var color = Chart.helpers.color;
+	var color = Chart.helpers.canvas.color;
 	var scatterChartData = {
 		datasets: [{
 			label: 'My First dataset',
