+++ samples/legend/point-style.html
@@ -35,7 +35,7 @@
 		</div>
 	</div>
 	<script>
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		function createConfig(colorName) {
 			return {
 				type: 'line',
