+++ samples/legend/positioning.html
@@ -41,7 +41,7 @@
 		</div>
 	</div>
 	<script>
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		function createConfig(legendPosition, colorName) {
 			return {
 				type: 'line',
