+++ samples/scales/logarithmic/scatter.html
@@ -19,7 +19,7 @@
 		<canvas id="canvas"></canvas>
 	</div>
 	<script>
-	var color = Chart.helpers.color;
+	var color = Chart.helpers.canvas.color;
 	var scatterChartData = {
 		datasets: [{
 			borderColor: window.chartColors.red,
