+++ samples/scales/time/combo.html
@@ -33,7 +33,7 @@
 			return moment().add(days, 'd').format(timeFormat);
 		}
 
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		var config = {
 			type: 'bar',
 			data: {
