+++ samples/scales/time/financial.html
@@ -57,7 +57,7 @@
 		ctx.canvas.width = 1000;
 		ctx.canvas.height = 300;
 
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		var cfg = {
 			type: 'bar',
 			data: {
