+++ samples/scales/time/line-point-data.html
@@ -33,7 +33,7 @@
 			return moment().add(days, 'd').format();
 		}
 
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		var config = {
 			type: 'line',
 			data: {
