+++ samples/scales/time/line.html
@@ -37,7 +37,7 @@
 			return moment().add(days, 'd').format(timeFormat);
 		}
 
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		var config = {
 			type: 'line',
 			data: {
