+++ samples/tooltips/custom-points.html
@@ -69,7 +69,7 @@
 				});
 			}
 		};
-		var color = Chart.helpers.color;
+		var color = Chart.helpers.canvas.color;
 		var lineChartData = {
 			labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
 			datasets: [{
