+++ src/core/core.animations.js
@@ -67,7 +67,7 @@ module.exports = {
 			// Skip animation frame requests until the active one is executed.
 			// This can happen when processing mouse events, e.g. 'mousemove'
 			// and 'mouseout' events will trigger multiple renders.
-			me.request = helpers.requestAnimFrame.call(window, function() {
+			me.request = helpers.canvas.requestAnimFrame.call(window, function() {
 				me.request = null;
 				me.startDigest();
 			});
